api/pyproject.toml
[tool.poetry]
name = "app"
version = "1.0.0"
description = "A platform for sharing and managing safely encoding asp"
authors = ["Pierfrancesco Napoli <pier.napoli@yahoo.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.112.1"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
annotated-types = "0.7.0"
anyio = "4.4.0"
bcrypt = "4.2.0"
black = "24.10.0"
certifi = "2024.7.4"
cffi = "1.17.1"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
colorlog = "6.8.2"
cryptography = "43.0.1"
deprecated = "1.2.14"
dnspython = "2.6.1"
email-validator = "2.2.0"
fastapi-cli = "0.0.5"
fastapi-pagination = "0.12.26"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
idna = "3.7"
importlib-resources = "6.4.4"
jinja2 = "3.1.4"
limits = "3.13.0"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
mdurl = "0.1.2"
mypy-extensions = "1.0.0"
packaging = "24.1"
passlib = "1.7.4"
pathspec = "0.12.1"
platformdirs = "4.3.6"
pyasn1 = "0.6.0"
pycparser = "2.22"
pydantic = "2.8.2"
pydantic-settings = "2.4.0"
pydantic-core = "2.20.1"
pygments = "2.18.0"
pyjwt = "2.9.0"
pymysql = "1.1.1"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
pyyaml = "6.0.2"
requests = "2.32.3"
rich = "13.7.1"
rsa = "4.9"
shellingham = "1.5.4"
slowapi = "0.1.9"
sniffio = "1.3.1"
sqlalchemy = "2.0.32"
starlette = "0.38.2"
typer = "0.12.4"
typing-extensions = "4.12.2"
urllib3 = "2.2.2"
uvloop = "0.21.0"
watchfiles = "0.23.0"
websockets = "13.0"
wrapt = "1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
foo='app.main:__main__'
